sms-number-hub-streamlit/requirements.txt
streamlit==1.41.1
pandas==2.2.3
python-dateutil==2.9.0.post0

fastapi==0.115.6
uvicorn[standard]==0.34.0
twilio==9.4.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
httpx==0.27.2
